Revert to Form from '@rjsf/core' for RegisterPage  (#84)
The Form component from '@rjsf/bootstrap-4' broke our bootstrap-3 styling and customized DescriptionField that renders HTML, causing the Lark reg form to fail to render. This fixes that by reverting to Form from '@rjsf/core' for now for the RegistrationPage.

Also includes a fix for a bug in the lark loadData script.