Prevent yarn.lock changing during container init
The way that Dockerfile and docker-compose.yml work together means that
when the container is built the yarn.lock could change if you just use
`yarn [install]`.  The --frozen-lockfile prevents any updates when the
container is first built.

After the build, docker-compose will run the container, mounting its
volumes, overwriting any changes made to the container during the build
phase.