Disable full lodging options (#176)
Adds code to count the number of registered campers for each lodging node and disable options that are full. Also appends "(full)" to the disabled option labels.

Redefines Lodging.reserved to mean the total number of reserved spots, rather than the remaining spots, and adds Camper.lodging_reserved as a way to indicate that the camper has been assigned to a reserved spot.

Adds documentation to lodging.py.

Adds a migration for previous changes to models.py.